requirements.txt
asgiref==3.8.1
Django==5.0.4
python-dotenv==1.0.1
sqlparse==0.5.0
django-phonenumber-field[phonenumberslite]==7.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

